requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiopg==1.4.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
ccxt==4.3.40
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
lxml==5.2.2
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
psycopg2-binary==2.9.9
pycares==4.4.0
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
SQLAlchemy==1.4.52
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
